+++ mapepire_python/client/websocket.py
@@ -4,7 +4,7 @@ from typing import Any, Dict
 
 from websocket import WebSocket, create_connection
 
-from ..types import DaemonServer
+from ..data_types import DaemonServer
 
 
 class WebsocketConnection:
@@ -29,7 +29,9 @@ class WebsocketConnection:
         return ssl_opts
 
     def connect(self) -> WebSocket:
-        try:  
-            return create_connection(self.uri, header=self.headers, sslopt=self.ssl_opts, timeout=10)
+        try:
+            return create_connection(
+                self.uri, header=self.headers, sslopt=self.ssl_opts, timeout=10
+            )
         except Exception as e:
-            raise Exception(f'An error occurred while connecting to the server: {e}')
+            raise Exception(f"An error occurred while connecting to the server: {e}")