+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 -- add cl tests
+ - add query manager
+
 
 ## [v0.1.4](https://github.com/Mapepire-IBMi/mapepire-python/releases/tag/v0.1.4) - 2024-08-23